refactor: rename authStore file
