+++ src/app/admin/page.tsx
@@ -8,7 +8,7 @@ import {useEffect, useState} from 'react'
 import {Modal} from '@/app/components'
 import {auth, db} from '@/libs/firebase'
 
-import {authStore} from '../auth/context/auth-state'
+import {authStore} from '../auth/context/auth-store'
 import {AddLinkForm, Header, LinksList} from './components'
 
 const Admin = observer(() => {
